simplify toggle logic
