+++ frontend/src/common/enums/comparisons/comparison-table-fields.ts
@@ -1,0 +1,12 @@
+enum TableFields {
+  bodyType = 'bodyType',
+  engine = 'engine',
+  engineDisplacement = 'engineDisplacement',
+  enginePower = 'enginePower',
+  colorName = 'colorName',
+  transmissionTypeName = 'transmissionTypeName',
+  drivetrainName = 'drivetrainName',
+  fuelTypeName = 'fuelTypeName',
+}
+
+export { TableFields };
